datasets/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
datasets==2.14.1
dill==0.3.7
filelock==3.12.2
frozenlist==1.4.0
fsspec==2023.6.0
huggingface-hub==0.16.4
idna==3.4
multidict==6.0.4
multiprocess==0.70.15
numpy==1.25.1
openai==0.27.8
packaging==23.1
pandas==2.0.3
pyarrow==12.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
tenacity==8.2.2
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
xxhash==3.2.0
yarl==1.9.2
